requirements.txt
Django<1.9
cassandra-driver>=2.5.0
